[hof] initial commit to support multiple offline finalized kernels
clamp-link would invoke hof for multiple times now, in order to produce
different kernels for different ISAs.

As HOF posted on Github doesn't support Fiji yet, clamp-link won't produce
kernels for Fiji as of now. Only Kaveri and Carrizo kernels would be produced.

In HCC runtime, each of these kernels would be checked to see if it's
compatible with the ISA on the HSA agent.

Currently the precedence is hard-coded as Fiji -> Kaveri -> Carrizo.